lunes 03

+++ src/main/java/GranPremio/Interfaces/Avanzable.java
@@ -1,6 +1,6 @@
 package GranPremio.Interfaces;
 
 public interface Avanzable {
-	double calcularAvanceTurno();
+	int calcularAvanceTurno();
 	void aplicarAvance(double metros);
 }
